style: reorder class members to match CODE_CONVENTION.md
Private readonly fields, then properties, then constructor, then
methods, applied repo-wide. No behavioral changes — build and full
test suite (1182/1182) verified green after the reorder.

+++ src/Moongate.Ultima/Animation/Animations.cs
@@ -527,7 +527,7 @@ public static class Animations
         int hue,
         bool isUop
     )
-        => (long)(body & 0xFFFFF) |
+        => body & 0xFFFFF |
            ((long)(action & 0x7F) << 20) |
            ((long)(direction & 0x7) << 27) |
            ((long)(fileType & 0x7) << 30) |
